# swagger.yaml
openapi: 3.0.0
info:
  title: Item API
  description: A simple API to manage items.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /items:
    get:
      summary: Get all items
      description: Retrieve a list of all items.
      responses:
        '200':
          description: A list of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
                  type: object
    post:
      summary: Create a new item
      description: Add a new item to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '201':
          description: Item created successfully.
  /items/{id}:
    get:
      summary: Get an item by ID
      description: Retrieve an item by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the item to retrieve.
      responses:
        '200':
          description: The requested item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found.
    put:
      summary: Update an item
      description: Update an item by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the item to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: Item updated successfully.
        '404':
          description: Item not found.
    delete:
      summary: Delete an item
      description: Remove an item by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the item to delete.
      responses:
        '204':
          description: Item deleted successfully.
        '404':
          description: Item not found.
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the item.
        name:
          type: string
          description: The name of the item.
        description:
          type: string
          description: The description of the item.
        price:
          type: number
          format: float
          description: The price of the item.
      required:
        - name
        - price